backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
bcrypt==4.1.2
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
redis==5.0.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
marshmallow==3.20.1
webargs==8.3.0
python-magic==0.4.27
python-magic-bin==0.4.14
clamd==1.0.2
user-agents==2.2.0